+++ datafusion/core/tests/sql/group_by.rs
@@ -442,3 +442,199 @@ async fn csv_group_by_date() -> Result<()> {
     assert_batches_sorted_eq!(expected, &actual);
     Ok(())
 }
+
+#[tokio::test]
+async fn group_by_date_trunc() -> Result<()> {
+    let tmp_dir = TempDir::new()?;
+    let ctx = SessionContext::new();
+    let schema = Arc::new(Schema::new(vec![
+        Field::new("c2", DataType::UInt64, false),
+        Field::new(
+            "t1",
+            DataType::Timestamp(TimeUnit::Microsecond, None),
+            false,
+        ),
+    ]));
+
+    // generate a partitioned file
+    for partition in 0..4 {
+        let filename = format!("partition-{}.{}", partition, "csv");
+        let file_path = tmp_dir.path().join(&filename);
+        let mut file = File::create(file_path)?;
+
+        // generate some data
+        for i in 0..10 {
+            let data = format!("{},2020-12-{}T00:00:00.000Z\n", i, i + 10);
+            file.write_all(data.as_bytes())?;
+        }
+    }
+
+    ctx.register_csv(
+        "test",
+        tmp_dir.path().to_str().unwrap(),
+        CsvReadOptions::new().schema(&schema).has_header(false),
+    )
+    .await?;
+
+    let results = plan_and_collect(
+        &ctx,
+        "SELECT date_trunc('week', t1) as week, SUM(c2) FROM test GROUP BY date_trunc('week', t1)",
+    ).await?;
+
+    let expected = vec![
+        "+---------------------+--------------+",
+        "| week                | SUM(test.c2) |",
+        "+---------------------+--------------+",
+        "| 2020-12-07 00:00:00 | 24           |",
+        "| 2020-12-14 00:00:00 | 156          |",
+        "+---------------------+--------------+",
+    ];
+    assert_batches_sorted_eq!(expected, &results);
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn group_by_largeutf8() {
+    {
+        let ctx = SessionContext::new();
+
+        // input data looks like:
+        // A, 1
+        // B, 2
+        // A, 2
+        // A, 4
+        // C, 1
+        // A, 1
+
+        let str_array: LargeStringArray = vec!["A", "B", "A", "A", "C", "A"]
+            .into_iter()
+            .map(Some)
+            .collect();
+        let str_array = Arc::new(str_array);
+
+        let val_array: Int64Array = vec![1, 2, 2, 4, 1, 1].into();
+        let val_array = Arc::new(val_array);
+
+        let schema = Arc::new(Schema::new(vec![
+            Field::new("str", str_array.data_type().clone(), false),
+            Field::new("val", val_array.data_type().clone(), false),
+        ]));
+
+        let batch =
+            RecordBatch::try_new(schema.clone(), vec![str_array, val_array]).unwrap();
+
+        let provider = MemTable::try_new(schema.clone(), vec![vec![batch]]).unwrap();
+        ctx.register_table("t", Arc::new(provider)).unwrap();
+
+        let results =
+            plan_and_collect(&ctx, "SELECT str, count(val) FROM t GROUP BY str")
+                .await
+                .expect("ran plan correctly");
+
+        let expected = vec![
+            "+-----+--------------+",
+            "| str | COUNT(t.val) |",
+            "+-----+--------------+",
+            "| A   | 4            |",
+            "| B   | 1            |",
+            "| C   | 1            |",
+            "+-----+--------------+",
+        ];
+        assert_batches_sorted_eq!(expected, &results);
+    }
+}
+
+#[tokio::test]
+async fn group_by_dictionary() {
+    async fn run_test_case<K: ArrowDictionaryKeyType>() {
+        let ctx = SessionContext::new();
+
+        // input data looks like:
+        // A, 1
+        // B, 2
+        // A, 2
+        // A, 4
+        // C, 1
+        // A, 1
+
+        let dict_array: DictionaryArray<K> =
+            vec!["A", "B", "A", "A", "C", "A"].into_iter().collect();
+        let dict_array = Arc::new(dict_array);
+
+        let val_array: Int64Array = vec![1, 2, 2, 4, 1, 1].into();
+        let val_array = Arc::new(val_array);
+
+        let schema = Arc::new(Schema::new(vec![
+            Field::new("dict", dict_array.data_type().clone(), false),
+            Field::new("val", val_array.data_type().clone(), false),
+        ]));
+
+        let batch =
+            RecordBatch::try_new(schema.clone(), vec![dict_array, val_array]).unwrap();
+
+        let provider = MemTable::try_new(schema.clone(), vec![vec![batch]]).unwrap();
+        ctx.register_table("t", Arc::new(provider)).unwrap();
+
+        let results =
+            plan_and_collect(&ctx, "SELECT dict, count(val) FROM t GROUP BY dict")
+                .await
+                .expect("ran plan correctly");
+
+        let expected = vec![
+            "+------+--------------+",
+            "| dict | COUNT(t.val) |",
+            "+------+--------------+",
+            "| A    | 4            |",
+            "| B    | 1            |",
+            "| C    | 1            |",
+            "+------+--------------+",
+        ];
+        assert_batches_sorted_eq!(expected, &results);
+
+        // Now, use dict as an aggregate
+        let results =
+            plan_and_collect(&ctx, "SELECT val, count(dict) FROM t GROUP BY val")
+                .await
+                .expect("ran plan correctly");
+
+        let expected = vec![
+            "+-----+---------------+",
+            "| val | COUNT(t.dict) |",
+            "+-----+---------------+",
+            "| 1   | 3             |",
+            "| 2   | 2             |",
+            "| 4   | 1             |",
+            "+-----+---------------+",
+        ];
+        assert_batches_sorted_eq!(expected, &results);
+
+        // Now, use dict as an aggregate
+        let results = plan_and_collect(
+            &ctx,
+            "SELECT val, count(distinct dict) FROM t GROUP BY val",
+        )
+        .await
+        .expect("ran plan correctly");
+
+        let expected = vec![
+            "+-----+------------------------+",
+            "| val | COUNT(DISTINCT t.dict) |",
+            "+-----+------------------------+",
+            "| 1   | 2                      |",
+            "| 2   | 2                      |",
+            "| 4   | 1                      |",
+            "+-----+------------------------+",
+        ];
+        assert_batches_sorted_eq!(expected, &results);
+    }
+
+    run_test_case::<Int8Type>().await;
+    run_test_case::<Int16Type>().await;
+    run_test_case::<Int32Type>().await;
+    run_test_case::<Int64Type>().await;
+    run_test_case::<UInt8Type>().await;
+    run_test_case::<UInt16Type>().await;
+    run_test_case::<UInt32Type>().await;
+    run_test_case::<UInt64Type>().await;
+}
