+++ datafusion/core/tests/sql/limit.rs
@@ -89,3 +89,77 @@ async fn csv_query_limit_zero() -> Result<()> {
     assert_batches_eq!(expected, &actual);
     Ok(())
 }
+
+#[tokio::test]
+async fn limit() -> Result<()> {
+    let tmp_dir = TempDir::new()?;
+    let ctx = create_ctx_with_partition(&tmp_dir, 1).await?;
+    ctx.register_table("t", table_with_sequence(1, 1000).unwrap())
+        .unwrap();
+
+    let results = plan_and_collect(&ctx, "SELECT i FROM t ORDER BY i DESC limit 3")
+        .await
+        .unwrap();
+
+    let expected = vec![
+        "+------+", "| i    |", "+------+", "| 1000 |", "| 999  |", "| 998  |",
+        "+------+",
+    ];
+
+    assert_batches_eq!(expected, &results);
+
+    let results = plan_and_collect(&ctx, "SELECT i FROM t ORDER BY i limit 3")
+        .await
+        .unwrap();
+
+    let expected = vec![
+        "+---+", "| i |", "+---+", "| 1 |", "| 2 |", "| 3 |", "+---+",
+    ];
+
+    assert_batches_eq!(expected, &results);
+
+    let results = plan_and_collect(&ctx, "SELECT i FROM t limit 3")
+        .await
+        .unwrap();
+
+    // the actual rows are not guaranteed, so only check the count (should be 3)
+    let num_rows: usize = results.into_iter().map(|b| b.num_rows()).sum();
+    assert_eq!(num_rows, 3);
+
+    Ok(())
+}
+
+#[tokio::test]
+async fn limit_multi_partitions() -> Result<()> {
+    let tmp_dir = TempDir::new()?;
+    let ctx = create_ctx_with_partition(&tmp_dir, 1).await?;
+
+    let partitions = vec![
+        vec![make_partition(0)],
+        vec![make_partition(1)],
+        vec![make_partition(2)],
+        vec![make_partition(3)],
+        vec![make_partition(4)],
+        vec![make_partition(5)],
+    ];
+    let schema = partitions[0][0].schema();
+    let provider = Arc::new(MemTable::try_new(schema, partitions).unwrap());
+
+    ctx.register_table("t", provider).unwrap();
+
+    // select all rows
+    let results = plan_and_collect(&ctx, "SELECT i FROM t").await.unwrap();
+
+    let num_rows: usize = results.into_iter().map(|b| b.num_rows()).sum();
+    assert_eq!(num_rows, 15);
+
+    for limit in 1..10 {
+        let query = format!("SELECT i FROM t limit {}", limit);
+        let results = plan_and_collect(&ctx, &query).await.unwrap();
+
+        let num_rows: usize = results.into_iter().map(|b| b.num_rows()).sum();
+        assert_eq!(num_rows, limit, "mismatch with query {}", query);
+    }
+
+    Ok(())
+}
